(Fixes #481) prevent e2e-test script from overriding task sidecar
The end to end test attaches a sidecar with a image registry for tasks to push into to. This task is optional and is encouraged to be used for testing. However, the function add_sidecar_registry sets the sidecar registry instead of appending it. Allowing the sidecar-registry to run alongside with other tasks will fix this bug. https://github.com/tektoncd/catalog/issues/481